+++ Graphics/MeshPoint.h
@@ -16,7 +16,7 @@ struct MeshPoint : public Point3D {
    Vector3D normal;
    bool texDefined;
    float texX, texY;
-   MeshPoint(float xIn = 0, float yIn = 0, float zIn = 0);
+   MeshPoint(double xIn = 0, double yIn = 0, double zIn = 0);
    void setTexPoint(float xIn, float yIn);
    void draw(bool drawNormal, bool drawTex);
 };
